requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bitsandbytes==0.43.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
datasets==2.20.0
dill==0.3.8
einops==0.8.0
filelock==3.15.4
flash-attn @ file:///home/jovyan/simple_rag/flash-attention
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.24.2
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
pluggy==1.5.0
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rouge-score==0.1.2
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
transformers-stream-generator==0.0.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.4.1
yarl==1.9.4
